requirements.txt
blessed==1.30.0
pyyaml==6.0.1
